pyproject.toml
[tool.poetry]
name = "paignation"
version = "0.1.0"
description = "페이지네이션 설명 프로젝트"
authors = ["heka1024 <heka1024@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-enumfields = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "^4.5.2"
black = "^21.12b0"
Faker = "^11.3.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
